+++ backend/base/src/models/takeaway_order.rs
@@ -8,11 +8,11 @@ pub struct Model {
     #[sea_orm(primary_key)]
     pub id: i32,
     pub uuid: Uuid,
-    pub train_order_id: i64,
-    pub takeaway_dish_id: i64,
-    pub person_info_id: i64,
-    pub pay_transaction_id: Option<i64>,
-    pub refund_transaction_id: Option<i64>,
+    pub train_order_id: i32,
+    pub takeaway_dish_id: i32,
+    pub person_info_id: i32,
+    pub pay_transaction_id: Option<i32>,
+    pub refund_transaction_id: Option<i32>,
     #[sea_orm(column_type = "Decimal(Some((10, 2)))")]
     pub price: Decimal,
     pub amount: i32,
